v4.0.0: rich and stability
