+++ aws_tower_lambda_launcher.py
@@ -26,9 +26,9 @@ from config import variables
 
 LOGGER = logging.getLogger('aws-tower-launcher')
 
-VERSION = '3.11.0'
+VERSION = '4.0.0'
 
-PATROWL = dict()
+PATROWL = {}
 PATROWL['api_token'] = os.environ['PATROWL_APITOKEN']
 PATROWL['assetgroup_dev'] = int(os.environ['PATROWL_DEV_ASSETGROUP'])
 PATROWL['assetgroup_pre'] = int(os.environ['PATROWL_PRE_ASSETGROUP'])
@@ -50,10 +50,10 @@ def organize_assetgroups(config):
     A Patrowl call is long and add latencies in every lambdas...
     """
     patrowl_assets = PATROWL_API.get_assets()
-    assetgroup = dict()
-    assetgroup['dev'] = list()
-    assetgroup['pre'] = list()
-    assetgroup['pro'] = list()
+    assetgroup = {}
+    assetgroup['dev'] = []
+    assetgroup['pre'] = []
+    assetgroup['pro'] = []
     for asset in patrowl_assets:
         for profile in config.sections():
             if not is_config_ok(config, profile):
