+++ libs/asset_type.py
@@ -9,6 +9,7 @@ Written by Nicolas BEGUIER (nicolas.beguier@adevinta.com)
 
 from dataclasses import dataclass
 
+from .tools import color_severity
 from .patterns import Patterns
 
 # Debug
@@ -57,23 +58,24 @@ class AssetType:
         """
         for issue in self.security_issues:
             if 'Audit' not in report:
-                report['Audit'] = list()
-            report['Audit'].append(f'[{issue["severity"]}] {issue["title"]}')
+                report['Audit'] = []
+            report['Audit'].append(color_severity(issue["severity"], issue["title"]))
 
     def display_brief_audit(self):
         """
         Return a brief output of the audit
         """
-        output = ''
-        if self.security_issues:
-            report = dict()
-            for issue in self.security_issues:
-                if issue['severity'] not in report:
-                    report[issue['severity']] = 1
-                else:
-                    report[issue['severity']] += 1
-            for issue_type in report:
-                output += f' [{issue_type}: {report[issue_type]}]'
+        if not self.security_issues:
+            return ''
+        output = ' '
+        report = {}
+        for issue in self.security_issues:
+            if issue['severity'] not in report:
+                report[issue['severity']] = 1
+            else:
+                report[issue['severity']] += 1
+        for severity, message in report.items():
+            output += f'<{color_severity(severity, message)}>'
         return output
 
     def get_type(self):
@@ -81,3 +83,11 @@ class AssetType:
         Return the asset type
         """
         return type(self).__name__
+
+    def remove_not_vulnerable_members(self):
+        """
+        If it's an AssetGroup (IAMGroup or S3Group or else), remove the
+        non vulnerable members.
+        This is a nutshell, check the ASsetGroup override function.
+        """
+        return True
