+++ libs/asset_type_rds.py
@@ -7,7 +7,11 @@ Licensed under the Apache License, Version 2.0
 Written by Nicolas BEGUIER (nicolas.beguier@adevinta.com)
 """
 
+# Third party library imports
+import botocore
+
 from .asset_type import AssetType
+from .tools import get_network
 
 # Debug
 # from pdb import set_trace as st
@@ -50,7 +54,7 @@ class RDS(AssetType):
         Return the report in one line
         """
         if self.public:
-            return f'[Public] {self.url} {self.engine}{self.display_brief_audit()}'
+            return f'<Public> {self.url} {self.engine}{self.display_brief_audit()}'
         return f'{self.engine}{self.display_brief_audit()}'
 
     def finding_description(self, _):
@@ -58,5 +62,53 @@ class RDS(AssetType):
         Return a description of the finding
         """
         if self.public:
-            return f'[Public] {self.url} {self.engine}'
-        return f'[Private] {self.engine}'
+            return f'<Public> {self.url} {self.engine}'
+        return f'<Private> {self.engine}'
+
+
+def get_raw_data(raw_data, authorizations, boto_session):
+    """
+    Get raw data from boto requests.
+    Return any RDS findings and add a 'False' in authorizations in case of errors
+    """
+    rds_client = boto_session.client('rds')
+    try:
+        raw_data['rds_raw'] = rds_client.describe_db_instances()['DBInstances']
+    except botocore.exceptions.ClientError:
+        raw_data['rds_raw'] = []
+        authorizations['rds'] = False
+    return raw_data, authorizations
+
+def scan(rds, subnets_raw, public_only):
+    """
+    Scan RDS
+    """
+    if public_only and not rds['PubliclyAccessible']:
+        return None
+    rds_asset = RDS(
+        name=rds['DBInstanceIdentifier'],
+        engine=f'{rds["Engine"]}=={rds["EngineVersion"]}',
+        public=rds['PubliclyAccessible'])
+    region, vpc, _ = get_network(
+        rds['DBSubnetGroup']['Subnets'][0]['SubnetIdentifier'],
+        subnets_raw)
+    rds_asset.location.region = region
+    rds_asset.location.vpc = vpc
+    rds_asset.location.subnet = rds['AvailabilityZone']
+    if 'Endpoint' in rds and 'Address' in rds['Endpoint']:
+        rds_asset.url = rds['Endpoint']['Address']
+    return rds_asset
+
+def parse_raw_data(assets, authorizations, raw_data, name_filter, public_only):
+    """
+    Parsing the raw data to extracts assets,
+    enrich the assets list and add a 'False' in authorizations in case of errors
+    """
+    for rds in raw_data['rds_raw']:
+        asset = scan(
+            rds,
+            raw_data['subnets_raw'],
+            public_only)
+        if asset is not None and name_filter.lower() in asset.name.lower():
+            assets.append(asset)
+    return assets, authorizations
