+++ libs/asset_type_s3_group.py
@@ -7,15 +7,14 @@ Licensed under the Apache License, Version 2.0
 Written by Nicolas BEGUIER (nicolas.beguier@adevinta.com)
 """
 
-# Standard library imports
-import logging
+# Third party library imports
+import botocore
 
 from .asset_type import AssetType
+from .asset_type_s3 import S3
 
 # Debug
 # from pdb import set_trace as st
-
-LOGGER = logging.getLogger('aws-tower')
 
 class S3Group(AssetType):
     """
@@ -71,3 +70,83 @@ class S3Group(AssetType):
             if iam.name == name:
                 return iam.finding_description(name)
         return 'S3 bucket not found...'
+
+    def remove_not_vulnerable_members(self):
+        """
+        Remove the non vulnerable members in the AssetGroup.
+        """
+        new_list = []
+        for s3_bucket in self.list:
+            if s3_bucket.security_issues:
+                new_list.append(s3_bucket)
+        self.list = new_list
+        return True
+
+def get_raw_data(raw_data, authorizations, boto_session):
+    """
+    Get raw data from boto requests.
+    Return any S3 findings and add a 'False' in authorizations in case of errors
+    """
+    s3_client = boto_session.client('s3')
+    raw_data['s3_client'] = s3_client
+    try:
+        raw_data['s3_list_buckets'] = s3_client.list_buckets()['Buckets']
+    except botocore.exceptions.ClientError:
+        raw_data['s3_list_buckets'] = []
+        authorizations['s3'] = False
+    return raw_data, authorizations
+
+def scan(s_three, configuration, region, acls, public_only):
+    """
+    Scan S3 Buckets
+    """
+    s3_asset = S3(
+        name=f's3://{s_three}',
+        url=f'https://{s_three}.s3.{region}.amazonaws.com/')
+    if configuration is None or not configuration['BlockPublicAcls']:
+        s3_asset.acls.block_public_acls = False
+    if configuration is None or not configuration['IgnorePublicAcls']:
+        s3_asset.acls.block_public_policy = False
+    if configuration is None or not configuration['BlockPublicPolicy']:
+        s3_asset.acls.ignore_public_acls = False
+    if configuration is None or not configuration['RestrictPublicBuckets']:
+        s3_asset.acls.restrict_public_buckets = False
+    s3_asset.update_grants(acls)
+    s3_asset.location.region = region
+    if public_only and not s3_asset.public:
+        return None
+    return s3_asset
+
+def parse_raw_data(assets, authorizations, raw_data, name_filter, public_only):
+    """
+    Parsing the raw data to extracts assets,
+    enrich the assets list and add a 'False' in authorizations in case of errors
+    """
+    s3group = S3Group(name='S3 buckets')
+    for s_three in raw_data['s3_list_buckets']:
+        try:
+            public_access_block_configuration = raw_data['s3_client'].get_public_access_block(
+                Bucket=s_three['Name'])['PublicAccessBlockConfiguration']
+        except botocore.exceptions.ClientError:
+            # Not really an error, the only way to know if it's a public bucket
+            public_access_block_configuration = None
+        try:
+            region = raw_data['s3_client'].get_bucket_location(Bucket=s_three['Name'])['LocationConstraint']
+        except botocore.exceptions.ClientError:
+            region = 'unknown'
+            authorizations['s3'] = False
+        try:
+            acls = raw_data['s3_client'].get_bucket_acl(Bucket=s_three['Name'])['Grants']
+        except botocore.exceptions.ClientError:
+            acls = []
+            authorizations['s3'] = False
+        s3bucket = scan(
+            s_three['Name'],
+            public_access_block_configuration,
+            region,
+            acls,
+            public_only)
+        if s3bucket is not None and name_filter.lower() in s3bucket.name.lower():
+            s3group.list.append(s3bucket)
+    assets.append(s3group)
+    return assets, authorizations
