+++ requirements.txt
@@ -1,2 +1,3 @@
 boto3>=1.16.63
 ruamel.yaml>=0.16.12
+rich>=10.12.0
